move StatusPages to appropriate package
